requirements.txt
# Core dependencies
agent-framework>=0.5.0
azure-identity>=1.15.0
openai>=1.12.0
pydantic>=2.5.0
aiohttp>=3.9.0
python-dotenv>=1.0.0

# Telemetry
opentelemetry-api>=1.22.0
opentelemetry-sdk>=1.22.0
opentelemetry-exporter-otlp>=1.22.0
opentelemetry-instrumentation-aiohttp-client>=0.43b0

# Persistence
redis>=5.0.0
asyncpg>=0.29.0

# MCP
mcp>=0.1.0

# Microsoft 365
msal>=1.26.0

# Utilities
tenacity>=8.2.0
structlog>=24.1.0
